src-torch1.6/requirements.txt
matplotlib==3.3.2
numpy==1.19.2
Pillow==8.0.1
scipy==1.5.3
seaborn==0.11.0
torch==1.6.0
torchvision==0.7.0
